+++ app/components/creators/shared/buttons/index.js
@@ -1,0 +1,1 @@
+export { default as JoinButton } from './JoinButton';
